# swagger.json
{
	"components": {
...
		"schemas": {
			"ERole": {
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"type": "string"
			},
			"EPermissions": {
				"enum": [
					"read",
					"write",
					"delete"
				],
				"type": "string"
			},
			"IPermissionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPermissions"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/ERole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"$ref": "#/components/schemas/IRoleDto"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserUpdateRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IUserListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IUserDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPrivilegesRequest": {
				"properties": {
					"roleName": {
						"$ref": "#/components/schemas/ERole"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/EPermissions"
						},
						"type": "array"
					}
				},
				"required": [
					"roleName",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IUserPassword": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/user/my": {
			"get": {
				"operationId": "GetMyUser",
				"responses": {
					"200": {
						"description": "Пользователь",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDto"
								}
							}
						}
					}
				},
				"description": "Получить пользователя.\nЭтот эндпоинт позволяет получить данные профиля пользователя, который выполнил запрос.\nИспользуется для получения информации о текущем пользователе, например, его имени, email, и других данных.",
				"summary": "Получение профиля текущего пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/my/update": {
			"patch": {
				"operationId": "UpdateMyUser",
				"responses": {
					"200": {
						"description": "Обновленный пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDto"
								}
							}
						}
					}
				},
				"description": "Обновить пользователя.\nЭтот эндпоинт позволяет пользователю обновить свои данные, такие как имя, email и другие параметры профиля.",
				"summary": "Обновление профиля текущего пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Обновленные данные профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateRequest",
								"description": "Обновленные данные профиля"
							}
						}
					}
				}
			}
		},
		"/api/user/my/delete": {
			"delete": {
				"operationId": "DeleteMyUser",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить текущего пользователя.\nЭтот эндпоинт позволяет удалить пользователя из системы.",
				"summary": "Удаление профиля текущего пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/list": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Список всех профилей с информацией о них",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserListDto"
								}
							}
						}
					}
				},
				"description": "Получить все профили.\nЭтот эндпоинт позволяет администраторам получить список всех пользователей системы.\nОн поддерживает пагинацию через параметры `offset` и `limit`.",
				"summary": "Получение всех профилей",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Смещение для пагинации",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Лимит количества возвращаемых профилей",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/user/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Пользователь по ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDto"
								}
							}
						}
					}
				},
				"description": "Получить пользователя по ID.\nЭтот эндпоинт позволяет получить пользователя по его ID. Доступен только для администраторов.",
				"summary": "Получение профиля по ID",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, которого нужно получить",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user/setPrivileges/{id}": {
			"patch": {
				"operationId": "SetPrivileges",
				"responses": {
					"200": {
						"description": "Обновленный пользователь с привилегиями",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDto"
								}
							}
						}
					}
				},
				"description": "Установить привилегии для пользователя.\nЭтот эндпоинт позволяет администраторам устанавливать роль и права пользователя.",
				"summary": "Установка привилегий для пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, для которого необходимо установить привилегии",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Запрос, содержащий роль и разрешения",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPrivilegesRequest",
								"description": "Запрос, содержащий роль и разрешения"
							}
						}
					}
				}
			}
		},
		"/api/user/requestVerifyEmail": {
			"post": {
				"operationId": "RequestVerifyEmail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Запросить подтверждение email-адреса для текущего пользователя.\nЭтот эндпоинт позволяет отправить пользователю письмо для подтверждения его email-адреса.",
				"summary": "Запрос подтверждения email",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/verifyEmail/{code}": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ответ о статусе подтверждения",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Подтвердить email-адрес текущего пользователя по коду.\nЭтот эндпоинт позволяет пользователю подтвердить свой email, используя код, полученный в письме.",
				"summary": "Подтверждение email-адреса",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Код подтверждения email, полученный пользователем",
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user/update/{id}": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Обновленный пользователь",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserDto"
								}
							}
						}
					}
				},
				"description": "Обновить пользователя.\nЭтот эндпоинт позволяет администраторам обновлять других пользователей.",
				"summary": "Обновление профиля другого пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, которого необходимо обновить",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Данные для обновления профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateRequest",
								"description": "Данные для обновления профиля"
							}
						}
					}
				}
			}
		},
		"/api/user/changePassword": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ответ о статусе изменения пароля",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Изменить пароль текущего пользователя.\nЭтот эндпоинт позволяет пользователю изменить свой пароль.",
				"summary": "Изменение пароля",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Новый пароль для пользователя",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPassword",
								"description": "Новый пароль для пользователя"
							}
						}
					}
				}
			}
		},
		"/api/user/delete/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить другого пользователя.\nЭтот эндпоинт позволяет администраторам удалить другого пользователя из системы.",
				"summary": "Удаление профиля другого пользователя",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, которого необходимо удалить",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}